Add components, remove line bug fix
Added a function to add given components of a given ion, fixed a
masking issue in plotting, renamed the unintuitive kwarg `line_labels`
to `legend` of plotting functions.
Fixed a bug in dataset.remove_line and added the keyword `levels` to
remove_fine_lines to match the functionality of add_fine_lines for
certain levels only.