Retain mask when adding line to region
When adding a line to a pre-existing region where a mask has been
defined (or altered, in case of global data-defined masks), the mask in
the existing region will now be propagated to the new region. In case
of conflicting masks between the global, data-defined mask (from
dataset.add_data) and the existing region’s mask, the existing region’s
mask has priority.